pyproject.toml
[build-system]
requires = [
    "setuptools >= 61",  # implement pyproject.toml compatibility
    "setuptools-scm[toml] >= 6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "escriba"
description = "Easy preservation of web content you care about."
readme = "README.rst"
license = {file = "COPYING"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Archiving",
]
requires-python = ">=3.11"
version = "0.0.1"
dependencies = [
    "aiosqlite",
    "flask",
    "pyzmq",
]

[project.optional-dependencies]
dev = [
    "black",
]

[tool.black]
line-length = 88
target-version = ['py311']
exclude = '''
(
  /(
      \.git          # exclude a few common directories in the
    | \.venv         # root of the project
    | \.mypy_cache
    | build
    | dist
    | docs
    | __pycache__
  )/
)
'''

[tool.setuptools_scm]
write_to = "escriba/_version.py"

